refactor(constant): isolate common string error
All messages output in the event of an exception start with "[ERROR]"
Add and use constant ERROR
