+++ src/main/java/lotto/constant/LottoGuide.java
@@ -11,7 +11,8 @@ public enum LottoGuide {
     RESULT_FOURTH("4개 일치 (50,000원) - %d개\n"),
     RESULT_THIRD("5개 일치 (1,500,000원) - %d개\n"),
     RESULT_SECOND("5개 일치, 보너스 볼 일치 (30,000,000원) - %d개\n"),
-    RESULT_FIRST("6개 일치 (2,000,000,000원) - %d개\n");
+    RESULT_FIRST("6개 일치 (2,000,000,000원) - %d개\n"),
+    ERROR("[ERROR] ");
 
     private final String message;
