+++ src/main/java/lotto/model/Consumer.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import lotto.constant.CompareInteger;
